implement login with User model
